first cut at supporting the "opaque" auth header param.
the idea and semantics are largely copied from http digest auth,
and is nicely described in rfc2617 as:

>   opaque
>     A string of data, specified by the server, which should be returned
>     by the client unchanged in the Authorization header of subsequent
>     requests with URIs in the same protection space. It is recommended
>     that this string be base64 or hexadecimal data.

this rfc is obsoleted by rfc7616, which doesn't really change the
semantics and use of opaque, it just takes more words to describe
it.

the idea is to mitigate the reuse of signature data between services,
or "protection spaces" on the same service. the current best
practice for that is to require that at least the "host" header and
"(request-target)" pseudo-header (and maybe the "(keyid)") get
included as input for the signature generation, but if you're doing
a lot of different requests for different URIs, you have to recalculate
the signature for every request against the same service. letting
the server provide an "opaque" value it can scope it's "protection
space" with, and then requiring "(opaque)" instead of "(request-target)"
as an input to the signature, allows the server to scope requests
as it sees fit, but allows the client to reuse signatures across
multiple requests.

if you're using something like a yubikey to sign requests, this is
a huge time saver.

discussed with alex wilson.