add a 'AuthSignatureOpaqueHandler' config parameter.
an operator can provide a module that provides an opaque value
programatically. if an opaque param is being checked, it will be
passed as an extra argument to the handler sub.

the handler has to return an Apache2::Const style status return.
if it returns DECLINED, opaque in a WWW-Authenticate header will
not be populated, and the value from an Authorization header won't
be checked. otherwise if it returns something other than OK, it
will cause the module to return that value to apache.